pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pubrecsearch"
version = "0.1.0"
description = "Scraping pipeline and search index for U.S. federal public records"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "polars>=0.20",
    "lxml>=5.0",
    "apscheduler>=3.10",
    "boto3>=1.34",
    "psycopg2-binary>=2.9",
    "tenacity>=8.2",
    "structlog>=24.1",
    "openpyxl>=3.1",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "python-dotenv>=1.0",
    "resend>=2.0",
    "pydantic-settings>=2.0",
    "click>=8.1",
    "truststore>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
]

[project.scripts]
pubrecsearch = "pubrecsearch.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/pubrecsearch"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
